+++ src/Umbraco.Web.UI.Client/src/packages/core/modal/types.ts
@@ -1,3 +1,6 @@
+import type { UUIModalElement, UUIModalSidebarSize } from '@umbraco-cms/backoffice/external/uui';
+import type { ElementLoaderProperty } from '@umbraco-cms/backoffice/extension-api';
+
 export type * from './extensions/index.js';
 
 export interface UmbPickerModalData<ItemType> {
@@ -14,4 +17,22 @@ export interface UmbPickerModalSearchConfig {
 
 export interface UmbPickerModalValue {
 	selection: Array<string | null>;
+}
+
+export type UmbModalType = 'dialog' | 'sidebar' | 'custom';
+
+export interface UmbModalConfig {
+	key?: string;
+	type?: UmbModalType;
+	size?: UUIModalSidebarSize;
+
+	/**
+	 * Used to provide a custom modal element to replace the default uui-modal-dialog or uui-modal-sidebar
+	 */
+	element?: ElementLoaderProperty<UUIModalElement>;
+
+	/**
+	 * Set the background property of the modal backdrop
+	 */
+	backdropBackground?: string;
 }
